made the sidebar an external thing
Because you put the sidebar in each file, it was hard to add a new entry in every file. And such, i added a template (i could not use jquery for it tho) which can be edited and linked in every new "chapter" (i put a description to the function that does that. i hope it is clear enough)